requirements.txt
streamlit>=1.30.0
python-dotenv>=1.0.0
google-genai>=1.0.0
requests>=2.31.0
python-whois>=0.9.4
pyyaml>=6.0
supabase>=2.0.0
